+++ src/main/java/io/github/astrapi69/resourcebundle/locale/LocaleResolver.java
@@ -36,14 +36,14 @@ import java.util.Set;
 import java.util.TreeSet;
 import java.util.stream.Collectors;
 
-import lombok.NonNull;
-import lombok.experimental.UtilityClass;
 import io.github.astrapi69.collection.list.ListExtensions;
 import io.github.astrapi69.collection.list.ListFactory;
 import io.github.astrapi69.io.file.FilenameExtensions;
 import io.github.astrapi69.lang.ClassExtensions;
 import io.github.astrapi69.resourcebundle.file.namefilter.PropertiesResourceBundleFilenameFilter;
 import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
+import lombok.NonNull;
+import lombok.experimental.UtilityClass;
 
 /**
  * The Class {@link LocaleResolver} helps to resolve locale objects and languages.