+++ src/main/java/io/github/astrapi69/resourcebundle/inspector/search/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides classes for the search and inspection of a properties files
+ */
+package io.github.astrapi69.resourcebundle.inspector.search;