+++ src/main/java/io/github/astrapi69/resourcebundle/inspector/validator/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides base classes for the validation and normalization of properties
+ */
+package io.github.astrapi69.resourcebundle.inspector.validator;