+++ src/main/java/io/github/astrapi69/resourcebundle/locale/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides bean and utility classes for locales and properties
+ */
+package io.github.astrapi69.resourcebundle.locale;